Develop 2

+++ src/room/room.service.ts
@@ -52,11 +52,11 @@ export class RoomService {
     }
 
     //change JSON Object to Array Object
-    const hashtags = createRoomDto.hashtag.toString()
+    const hashtags = createRoomDto.hashtag
+      .toString()
       .replace(/\[|\]/g, '')
       .replace(/\s/g, '')
       .split(',');
-
 
     //Need to delete this part after putting res for res.locals.userid
     const userId = ['dopeDude'];
@@ -83,6 +83,7 @@ export class RoomService {
           content: tag,
           rooms: [result.id as string],
         });
+
         await newHashtag.save();
       }
 
@@ -140,11 +141,12 @@ export class RoomService {
   //Need to think what can go wrong with this method. - ***CreatedAt has changed, If Image is not provided it will become default image
   async updateRoom(file, roomId, updateRoomDto) {
     const targetRoom = await this.findRoom(roomId);
-    const filename = targetRoom.imageLocation;
+    let filename = targetRoom.imageLocation;
+    let hashtags = targetRoom.hashtags;
 
     //if new image is provided, delete original image
     if (file) {
-      const filename = file.filename;
+      filename = file.filename;
       if (targetRoom.imageLocation != 'defaultImage.png') {
         await fs.unlink(
           `./public/roomImages/${targetRoom.imageLocation}`,
@@ -159,33 +161,44 @@ export class RoomService {
     }
 
     //reset Hashtags
-    for (let i = 0; i < targetRoom.hashtags.length; i++) {
-      const tag = targetRoom.hashtags[i];
-      const dbHashtag = await this.findHashtag(tag);
+    if (updateRoomDto.hashtag) {
+      for (let i = 0; i < targetRoom.hashtags.length; i++) {
+        const tag = targetRoom.hashtags[i];
+        const dbHashtag = await this.findHashtag(tag);
 
-      //If hashtag length is 1, delete hashtag from DB, else, remove roomId from hashtag.rooms
-      if (dbHashtag.rooms.length == 1) {
-        await this.hashtagModel.deleteOne({ _id: dbHashtag._id }).exec();
-      } else {
-        await this.hashtagModel.updateOne(
-          { _id: dbHashtag._id },
-          { $pull: { rooms: roomId } },
-        );
+        //If hashtag length is 1, delete hashtag from DB, else, remove roomId from hashtag.rooms
+
+        if (dbHashtag.rooms.length == 1) {
+          await this.hashtagModel.deleteOne({ _id: dbHashtag._id }).exec();
+        } else {
+          await this.hashtagModel.updateOne(
+            { _id: dbHashtag._id },
+            { $pull: { rooms: roomId } },
+          );
+        }
       }
-    }
 
-    //recreate Hashtags
-    for (let i = 0; i < updateRoomDto.hashtags.length; i++) {
-      const tag = updateRoomDto.hashtags[i];
-      const dbHashtag = await this.findHashtag(tag);
+      //change JSON Object to Array Object
 
-      //if no hashtag exists, create one
-      if (!dbHashtag) {
-        const newHashtag = new this.hashtagModel({
-          content: tag,
-          rooms: [roomId as string],
-        });
-        await newHashtag.save();
+      hashtags = updateRoomDto.hashtag
+        .toString()
+        .replace(/\[|\]/g, '')
+        .replace(/\s/g, '')
+        .split(',');
+
+      //recreate Hashtags
+      for (let i = 0; i < hashtags.length; i++) {
+        const tag = hashtags[i];
+        const dbHashtag = await this.findHashtag(tag);
+
+        //if no hashtag exists, create one
+        if (!dbHashtag) {
+          const newHashtag = new this.hashtagModel({
+            content: tag,
+            rooms: [roomId as string],
+          });
+          await newHashtag.save();
+        }
       }
     }
 
@@ -194,6 +207,7 @@ export class RoomService {
       {
         $set: {
           ...updateRoomDto,
+          hashtags,
           imageLocation: filename,
           lastVisited: new Date(),
         },
